[MAT-425] compile under stricter rulings
This changeset does a large fixup sweep to get the code to
compile under -std=gnu99. gnu99 is used in preference to c99 to
allow access to the GNU extensions like `M_PI` in `math.h`.

Also in this changeset is the addition of cmake rules to enable
multilib support as per the rest of the OG-Maths chain. The code
is added but the wiring isn't working end to end yet.