[MAT-425] Add in cmake export rules
The cmake export rules are added such that the key targets
made available by this library are directly exportable from the
build tree. A minor hack is applied to export the header file
location in the export file too.

The "izy.h" header is also modified to be C++ compatible.